+++ repeat-test.sh
@@ -2,6 +2,7 @@
 for i in {1..1000}
 do
    echo "TEST #$i"
-   npm test
+   # npm test
+   ./test/single_test.sh
    echo "TEST #$i end."
 done